Factor out Unix and WASI fd code into a common module.
